+++ app/graphql/ssbm_data_schema.rb
@@ -1,5 +1,0 @@
-class SsbmDataSchema < GraphQL::Schema
-  # mutation(Types::MutationType)
-  # character(Types::CharacterType)
-  query(Types::QueryType)
-end